# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Callception API",
    "description": "AI-powered call management, appointment tracking, and customer service automation platform. Provides voice AI pipeline, CRM, ticketing, billing, and multi-tenant management capabilities.",
    "version": "1.0.0",
    "contact": {
      "name": "Callception",
      "url": "https://callception.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://callception.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3002",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "TenantHeader": []
    }
  ],
...
  "paths": {
...
    "/api/voice/tts": {
      "post": {
        "tags": ["Voice Pipeline"],
        "summary": "Text-to-Speech synthesis",
        "description": "Converts text to audio using Cartesia Sonic-3 (primary) with multi-provider fallback. Returns audio/mpeg stream.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string" },
                  "voice_id": { "type": "string", "description": "TTS voice ID override" },
                  "language": { "type": "string", "enum": ["tr", "en"] },
                  "greeting": { "type": "boolean", "default": false },
                  "provider": { "type": "string", "enum": ["cartesia", "murf", "kokoro", "google", "openai"] },
                  "session_id": { "type": "string" },
                  "tenant_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audio stream",
            "content": { "audio/mpeg": { "schema": { "type": "string", "format": "binary" } } },
            "headers": {
              "X-TTS-Provider": { "schema": { "type": "string" } },
              "X-TTS-Latency-Ms": { "schema": { "type": "string" } },
              "X-TTS-Language": { "schema": { "type": "string" } }
            }
          },
          "503": { "description": "All TTS providers failed" }
        }
      },
      "get": {
        "tags": ["Voice Pipeline"],
        "summary": "TTS provider status",
        "responses": { "200": { "description": "Provider configuration and performance info" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase JWT",
        "description": "Firebase ID token for authenticated requests"
      },
      "TenantHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-tenant",
        "description": "Tenant ID header set by authentication middleware"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for automation/webhook endpoints"
      }
    },
...
  }
}